no more OOP
deleted class stuff and now it's all one giant main method
