--- treegen/consts.h
@@ -1,12 +1,0 @@
-#ifndef CONSTS_H
-#define CONSTS_H
-
-#include <iostream>
-using namespace std;
-
-static const int SPECIES = 10;
-static const double P_MUTATE = 0.2;
-static const int SEQUENCE_LENGTH = 1000;
-static const string SPECIES_NAMES[SPECIES] = {"A", "B", "C", "D", "E", "F", "G", "H", "I", "J"};
-
-#endif
